requirements.txt
# pip install -r requirements.txt
Django==4.1.6
djangorestframework==3.14.0
django-summernote==0.8.20.0
Pillow==9.4.0